requirements.txt
fastapi==0.131.0
uvicorn==0.41.0
paddleocr==3.4.0
paddlepaddle-gpu==3.3.0
openai==2.21.0
sqlalchemy==2.0.46
jieba==0.42.1
pillow==12.1.0
opencv-contrib-python==4.10.0.84
pydantic==2.12.5
python-dotenv==1.2.1
python-multipart==0.0.22
aiofiles>=23.0.0
httpx>=0.24.0
zai-sdk>=0.1.0
